requirements.txt
openai==1.30.1
httpx==0.27.0
fastapi==0.115.5
uvicorn==0.32.1
python-telegram-bot==22.5
SQLAlchemy==2.0.36
aiosqlite==0.22.1
python-dotenv==1.0.0
edge-tts==7.2.7
redis[asyncio]==5.2.1
Pillow>=10.0
pytest==8.3.4
pytest-asyncio==0.24.0
